requirements.txt
affine==2.4.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
cycler==0.11.0
earthpy==0.9.4
Fiona==1.8.22
fonttools==4.38.0
geopandas==0.12.2
idna==3.4
imageio==2.25.0
kiwisolver==1.4.4
matplotlib==3.6.3
munch==2.5.0
networkx==3.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
rasterio==1.3.5
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.0
shapely==2.0.0
six==1.16.0
snuggs==1.4.7
tifffile==2023.1.23.1
urllib3==1.26.14
